(MODULES-7051) re rename
Prior to this commit ...

References to `puppet_device` were globally replaced with 'device_manager'
resulting in unanticipated changes.

With this commit ...

The custom fact is renamed from
`device_managers` to `devices`.

The task that executes a `puppet device` run is renamed
from `device_manager` to `device_manager::run_puppet_device`.

The namespace of the task is renamed
from `tkishel/device_manager` to `puppetlabs/device_manager`.

File name variables are normalized.

Resource tags and titles are normalized.

Example values are normalized.

The README is updated with these same changes.