(WIP) Specify default values when using Hiera or Classifier.
This commit corrects the merge_default and device_manager::devices,
adds spec tests for the merge_default and value_or_default functions,
and simplifes the username/password in the examples.

TODO: Update the README to document default values,
the three tiers for values: global default, type default, and parameter,
and their precedence.