(FM-7383) allow device manager to work with Puppet 6
Puppet 6.0.0 removed `puppet cert` and it is now part of `puppetserver
ca`. This PR changes the behaviour to use `puppetserver ca sign/list`
when running on Puppet 6.0.0 and above.

ssl_host suitable algorithms array was also removed, so referencing the
array directly to overcome this behaviour as no mention in Puppet 6.0.0
as to where it was moved too.