(MODULES-9628, PUP-8736) don't purge device certificates on new puppet versions
The changes in PUP-8736 were causing the recurse/purge options here
to delete certificates that should remain persisted. This hotfix
prevents the data loss by only purging configuration files at the
top-level, but not certificates in subdirectories.

Additionally, this changes the owner/group of the `devices` directory
to use the puppetmaster's user/group settings to line up with what
puppet itself forces the permissions of the directory to be.